+++ DfE.GIAP.All/tests/DfE.GIAP.Core.UnitTests/MyPupils/Infrastructure/Repository/CosmosDbUserReadOnlyRepositoryTests.cs
@@ -12,27 +12,18 @@ using Microsoft.Azure.Cosmos;
 namespace DfE.GIAP.Core.UnitTests.MyPupils.Infrastructure.Repository;
 public sealed class CosmosDbUserReadOnlyRepositoryTests
 {
-
-    [Fact]
-    public void Constructor_ThrowsArgumentNullException_When_CommandHandlerIsNull()
-    {
-        // Arrange
-        Func<CosmosDbUserWriteRepository> construct = () => new CosmosDbUserWriteRepository(
-            commandHandler: null!,
-            logger: LoggerTestDoubles.MockLogger<CosmosDbUserWriteRepository>());
-
-        // Act Assert
-        Assert.Throws<ArgumentNullException>(construct);
-    }
-
-
     [Fact]
     public void Constructor_ThrowsArgumentNullException_When_LoggerIsNull()
     {
         // Arrange
-        Func<CosmosDbUserWriteRepository> construct = () => new CosmosDbUserWriteRepository(
-            commandHandler: CosmosDbCommandHandlerTestDoubles.Default().Object,
-            logger: null!);
+        Mock<IMapper<UserDto, User.Application.User>> mockMapper = MapperTestDoubles.Default<UserDto, User.Application.User>();
+
+        Mock<ICosmosDbQueryHandler> mockCosmosDbQueryHandler = CosmosDbQueryHandlerTestDoubles.Default();
+
+        Func<CosmosDbUserReadOnlyRepository> construct = () => new CosmosDbUserReadOnlyRepository(
+            logger: null!,
+            cosmosDbQueryHandler: mockCosmosDbQueryHandler.Object,
+            userMapper: mockMapper.Object);
 
         // Act Assert
         Assert.Throws<ArgumentNullException>(construct);
