+++ DfE.GIAP.All/tests/DfE.GIAP.Core.UnitTests/MyPupils/Infrastructure/Repository/CosmosDbUserWriteRepositoryTests.cs
@@ -1,9 +1,10 @@
 ﻿using Dfe.Data.Common.Infrastructure.Persistence.CosmosDb.Handlers.Command;
 using DfE.GIAP.Core.MyPupils.Domain.ValueObjects;
-using DfE.GIAP.Core.User.Application;
-using DfE.GIAP.Core.User.Infrastructure.Repository.Dtos;
-using DfE.GIAP.Core.User.Infrastructure.Repository;
 using DfE.GIAP.Core.SharedTests.TestDoubles;
+using DfE.GIAP.Core.UnitTests.TestDoubles;
+using DfE.GIAP.Core.User.Application;
+using DfE.GIAP.Core.User.Infrastructure.Repository;
+using DfE.GIAP.Core.User.Infrastructure.Repository.Dtos;
 using DfE.GIAP.SharedTests.TestDoubles;
 using Microsoft.Azure.Cosmos;
 
@@ -11,6 +12,31 @@ namespace DfE.GIAP.Core.UnitTests.MyPupils.Infrastructure.Repository;
 public sealed class CosmosDbUserWriteRepositoryTests
 {
     private const string UsersContainerName = "users";
+
+    [Fact]
+    public void Constructor_ThrowsArgumentNullException_When_CommandHandlerIsNull()
+    {
+        // Arrange
+        Func<CosmosDbUserWriteRepository> construct = () => new CosmosDbUserWriteRepository(
+            commandHandler: null!,
+            logger: LoggerTestDoubles.MockLogger<CosmosDbUserWriteRepository>());
+
+        // Act Assert
+        Assert.Throws<ArgumentNullException>(construct);
+    }
+
+
+    [Fact]
+    public void Constructor_ThrowsArgumentNullException_When_LoggerIsNull()
+    {
+        // Arrange
+        Func<CosmosDbUserWriteRepository> construct = () => new CosmosDbUserWriteRepository(
+            commandHandler: CosmosDbCommandHandlerTestDoubles.Default().Object,
+            logger: null!);
+
+        // Act Assert
+        Assert.Throws<ArgumentNullException>(construct);
+    }
 
     [Fact]
     public async Task SaveMyPupilsAsync_MapsUpnsAndCallsUpsert_WithExpectedDto()
